GUI in einzelne Klassen aufgeteilt
